refactor: move `unknown_length` to new submodule
